+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+- Implement find protocols and multimethods in project
+
 ## 0.1.12 (2019-12-12)
 
 - Upgrade to clindex 0.4.1